requirements.txt
pytest
pandas 
scipy
numpy 
scikit-learn == 1.0.2
joblib
uvicorn
fastapi
pydantic
wandb
keras
tensorflow-cpu